// interfaces/finance.interface.ts
export interface IFinance {
  publishDate?: string;
  description?: string;
  marketGroup?: string;
  modelYear?: string;
  namePlate?: string;
  bookCode?: string;
  vehicleLine?: string;
  merchandiseModelCode?: string;
  restrictionString?: string;
  fromDate?: string;
  toDate?: string;
  fromPublishId?: string;
  toPublishId?: string;
  mmcCodeSP?: string;
  mmcCode?: string;
  modelRestrictionString?: string;
  comparisionReportType?: string;
  publishId?: any;
}